pyproject.toml
[project]
name = "demo-ping-pong"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bottle>=0.13.4",
    "pycryptodome>=3.23.0",
]
